+++ application/models/Conf.php
@@ -30,7 +30,7 @@ class Conf extends CI_Model
         setRedirect(); //remember current page
 
         //enable profiler in dev
-        if(ENVIRONMENT == 'development') {
+        if(ENVIRONMENT == 'testing') {
             $this->output->enable_profiler(TRUE);
 //            $this->console->exception(new Exception('test exception'));
 //            $this->console->debug('Debug message');